xcvbn

+++ src/devices/devices.service.ts
@@ -13,6 +13,7 @@ import { Device } from './Schema/Device'
 import { CreateDeviceDto } from './dto/create-device.dto'
 import { DeviceQueryDto } from './dto/device.query.dto'
 import { UpdateDeviceDto } from './dto/update-device.dto'
+import { passport } from 'src/_shared'
 
 @Injectable()
 export class DevicesService {
@@ -20,30 +21,26 @@ export class DevicesService {
 
   async create (createDeviceDto: CreateDeviceDto, file: Express.Multer.File) {
     try {
+      const data = xlsxToArray(file?.path)
+      if (!data) {
+        file?.filename && deleteFile('uploads', file.filename)
+        throw new BadRequestException({ msg: 'Invalid xlsx file ' })
+      }
       const dpkExist = await this.deviceModel.findOne({
-        device_privet_key: createDeviceDto.device_privet_key,
+        $or: [
+          { device_privet_key: createDeviceDto.device_privet_key },
+          { serie: createDeviceDto.serie },
+        ],
       })
-      const serieExist = await this.deviceModel.findOne({
-        serie: createDeviceDto.serie,
-      })
-      if (serieExist || dpkExist) {
-        deleteFile('uploads', file.filename)
+      if (dpkExist) {
+        file.filename && deleteFile('uploads', file.filename)
         throw new BadRequestException({
-          msg: `${serieExist ? serieExist.serie : ''} ${
-            dpkExist ? dpkExist.serie : ''
-          }  already exists!`,
+          msg: `Device private key or serie already exists!`,
+          exist: dpkExist,
         })
       }
 
-      const data = xlsxToArray(file.path)
-      if (data.length === 0) {
-        deleteFile('uploads', file.filename)
-        throw new BadRequestException({ msg: 'Invalid xlsx file ' })
-      }
-      write(
-        `./passports/${createDeviceDto.serie}.json`,
-        convertArrayToJSON(data)
-      )
+      write(`./passports/${createDeviceDto.serie}.json`, data as any[])
       return this.deviceModel.create(createDeviceDto)
     } catch (error) {
       throw new BadRequestException({
@@ -85,7 +82,10 @@ export class DevicesService {
       const data = await this.deviceModel.find(filter)
       return { data, limit: 0, offset: 0, total }
     } catch (error) {
-      throw new BadRequestException({ msg: "Keyinroq urinib ko'ring..." , error })
+      throw new BadRequestException({
+        msg: "Keyinroq urinib ko'ring...",
+        error,
+      })
     }
   }
 
@@ -104,7 +104,10 @@ export class DevicesService {
         .skip(limit * offset)
       return { data, limit, offset, total }
     } catch (error) {
-      throw new BadRequestException({ msg: "Keyinroq urinib ko'ring..." , error })
+      throw new BadRequestException({
+        msg: "Keyinroq urinib ko'ring...",
+        error,
+      })
     }
   }
 
@@ -115,7 +118,10 @@ export class DevicesService {
         { path: 'owner', select: 'username first_name last_name' },
       ])
     } catch (error) {
-      throw new BadRequestException({ msg: "Keyinroq urinib ko'ring..." , error })
+      throw new BadRequestException({
+        msg: "Keyinroq urinib ko'ring...",
+        error,
+      })
     }
   }
 
@@ -126,50 +132,54 @@ export class DevicesService {
   ) {
     try {
       const exist = await this.deviceModel.findById(id)
-    if (!exist) {
-      deleteFile('uploads', file?.filename)
-      throw new BadRequestException({ msg: 'Device does not exist!' })
-    }
-    const dpkExist = await this.deviceModel.findOne({
-      device_privet_key: updateDeviceDto.device_privet_key,
-      _id: { $ne: id },
-    })
-    const serieExist = await this.deviceModel.findOne({
-      serie: updateDeviceDto.serie,
-      _id: { $ne: id },
-    })
-    if (serieExist && dpkExist) {
-      file?.filename && deleteFile('uploads', file.filename)
+      if (!exist) {
+        deleteFile('uploads', file?.filename)
+        throw new BadRequestException({ msg: 'Device does not exist!' })
+      }
+      const dpkExist = await this.deviceModel.findOne({
+        device_privet_key: updateDeviceDto.device_privet_key,
+        _id: { $ne: id },
+      })
+      const serieExist = await this.deviceModel.findOne({
+        serie: updateDeviceDto.serie,
+        _id: { $ne: id },
+      })
+      if (serieExist || dpkExist) {
+        file?.filename && deleteFile('uploads', file.filename)
+        throw new BadRequestException({
+          msg: 'Device private key or serie already exists!',
+        })
+      }
+      if (file?.path && updateDeviceDto.serie) {
+        const data = xlsxToArray(file?.path)
+        if (!data) {
+          file?.filename && deleteFile('uploads', file?.filename)
+          throw new BadRequestException({ msg: 'Invalid xlsx file ' })
+        }
+        deleteFile('passports', `${exist.serie}.json`)
+        file?.filename &&
+          write(
+            `./passports/${updateDeviceDto.serie || exist.serie}.json`,
+            data as any
+          )
+      }
+
+      const updated = await this.deviceModel.findByIdAndUpdate(
+        id,
+        updateDeviceDto,
+        { new: true }
+      )
+      if (updated) {
+        return { msg: 'Qurilma yangilandi.' }
+      } else {
+        throw new BadRequestException({ msg: 'Qurilmani yangilashda xatolik' })
+      }
+    } catch (error) {
       throw new BadRequestException({
-        msg: 'Device private key or serie already exists!',
+        msg: "Keyinroq urinib ko'ring...",
+        error,
       })
     }
-    const data = xlsxToArray(file?.path)
-    if (data.length === 0 && file?.path) {
-      file?.filename && deleteFile('uploads', file?.filename)
-      throw new BadRequestException({ msg: 'Invalid xlsx file ' })
-    }
-
-    file?.filename &&
-      write(
-        `./passports/${updateDeviceDto.serie || exist.serie}.json`,
-        convertArrayToJSON(data)
-      )
-    deleteFile('passports', `${exist.serie}.json`)
-    const updated = await this.deviceModel.findByIdAndUpdate(
-      id,
-      updateDeviceDto,
-      { new: true }
-    )
-    if (updated) {
-      return { msg: 'Qurilma yangilandi.' }
-    } else {
-      throw new BadRequestException({ msg: 'Qurilmani yangilashda xatolik' })
-    }
-    } catch (error) {
-      throw new BadRequestException({ msg: "Keyinroq urinib ko'ring..." , error })
-    }
-    
   }
 
   async remove ({ id }: ParamIdDto) {
@@ -183,7 +193,10 @@ export class DevicesService {
         throw new BadRequestException({ msg: 'Qurilmani ochirishda xatolik' })
       }
     } catch (error) {
-      throw new BadRequestException({ msg: "Keyinroq urinib ko'ring..." , error })
+      throw new BadRequestException({
+        msg: "Keyinroq urinib ko'ring...",
+        error,
+      })
     }
   }
 }
